ImageAzureML/requirements.txt
appdirs==1.4.2
packaging==16.8
pyparsing==2.2.0
six==1.10.0
Flask==0.12
pylint==1.6.4
numpy==1.11.2
pandas==0.19.0
Pillow==3.4.2